Updated RecipeID to Long

+++ backend/src/main/java/recipes/fridger/backend/service/RecipeServiceImpl.java
@@ -30,7 +30,7 @@ public class RecipeServiceImpl implements RecipeService{
     }
     @Transactional
     @Override
-    public void deleteRecipe(Integer id) {
+    public void deleteRecipe(Long id) {
         Optional<Recipe> recipe = recipes.findById(id);
         if (recipe.isPresent()) {
             Recipe r = recipe.get();
@@ -39,7 +39,7 @@ public class RecipeServiceImpl implements RecipeService{
     }
     @Transactional
     @Override
-    public Recipe getRecipe(Integer id) {
+    public Recipe getRecipe(Long id) {
         Optional<Recipe> recipe = recipes.findById(id);
         if (recipe.isPresent()) {
             return recipe.get();
@@ -47,7 +47,7 @@ public class RecipeServiceImpl implements RecipeService{
         return null;
     }
     @Transactional
-    public Iterable<Recipe> getRecipes(Integer userId) {
+    public Iterable<Recipe> getRecipes(Long userId) {
         return recipes.find(userId);
     }
 }
